refactor(#6712): core/interfaces 整层下架 + ML 策略三变体统一
ADR-022 原则6 第二期：命名空间唯一性。

删除（0 生产引用，grep 复核）:
- src/ginkgo/core/interfaces/{__init__,strategy_interface,portfolio_interface}.py
  (BaseStrategy/BasePortfolio core 版 + BaseMLStrategy/BaseMultiStrategyPortfolio)
- src/ginkgo/quant_ml/strategies/ml_strategy_base.py (quant_ml 域 MLStrategyBase 死类)
- core/interfaces/model_interface.py 的死子类 BaseTimeSeriesModel/BaseEnsembleModel
- tests/unit/core/interfaces/test_{strategy,portfolio,model}_interface.py
- core/__init__.py 的 interfaces try/except + __all__ 入口

迁移:
- BaseModel + ModelStatus: core/interfaces/model_interface.py -> quant_ml/models/base_model.py
  (5 处生产 import 全在 quant_ml 域: lightgbm/sklearn/xgboost + 2 个 ML 策略基类)
- ML 策略统一为 MLStrategyBase @ trading/strategies/ml_strategy_base.py
  (原 StrategyMLBase 改名; quant_ml 版删除; PredictionStrategy 改继承 trading 版,
   自带 Bar 事件管线以保原行为; core 版 BaseMLStrategy 随 core 下架)

护栏:
- 扩展 tests/unit/interfaces/test_dead_abstract_removed.py: 断言 core.interfaces
  包及 strategy/portfolio/model 子模块均不可 import; BaseModel 已迁 quant_ml;
  MLStrategyBase 统一归 trading

冒烟三层全绿:
- L1 py_compile (src+api): 0 错
- L2 user_crud_mock/containers/user_cli: 36/37 (1 pre-existing 失败, master baseline 同样红,
  click rich 渲染把 --name 拆成 ANSI 序列, 与本 PR 无关)
- L3 test_dead_abstract_removed + test_ml_strategy + test_containers: 43/43

Closes #6712
