+++ src/ginkgo/quant_ml/models/sklearn.py
@@ -30,7 +30,7 @@ try:
 except ImportError:
     SKLEARN_AVAILABLE = False
 
-from ginkgo.core.interfaces.model_interface import BaseModel, ModelStatus
+from ginkgo.quant_ml.models.base_model import BaseModel, ModelStatus
 from ginkgo.enums import MODEL_TYPES
 from ginkgo.libs import GLOG
