+++ src/ginkgo/trading/strategies/__init__.py
@@ -1,5 +1,5 @@
 # Upstream: EngineAssemblyService, ComponentLoader, PortfolioBase
-# Downstream: BaseStrategy, RandomSignalStrategy, StrategyMLBase, StrategyMLPredictor
+# Downstream: BaseStrategy, RandomSignalStrategy, MLStrategyBase, StrategyMLPredictor
 # Role: 策略模块包，导出策略基类及ML策略等策略组件
 #
 # ===== 组件边界 =====
@@ -32,12 +32,12 @@ from ginkgo.trading.strategies.momentum import Momentum
 
 # Machine Learning Strategies
 try:
-    from ginkgo.trading.strategies.ml_strategy_base import StrategyMLBase
+    from ginkgo.trading.strategies.ml_strategy_base import MLStrategyBase
     from ginkgo.trading.strategies.ml_predictor import StrategyMLPredictor
     ML_STRATEGIES_AVAILABLE = True
 except ImportError as e:
     # ML strategies are optional and require additional dependencies
     ML_STRATEGIES_AVAILABLE = False
-    StrategyMLBase = None
+    MLStrategyBase = None
     StrategyMLPredictor = None
